Move openModal inside image gallery
