+++ src/pages/hobbies/Lego.jsx
@@ -3,14 +3,6 @@ import React, { useState } from "react";
 import ImageGallery from "../components/ImageGallery";
 
 function Lego() {
-  const [image, setImage] = useState(null);
-  const [caption, setCaption] = useState("");
-
-  const openModal = (image, caption) => {
-    setImage(image);
-    setCaption(caption);
-  };
-
   // image imports
   function importAll(r) {
     return r.keys().map(r);
@@ -32,22 +24,7 @@ function Lego() {
           medieval encampment.
         </p>
       </div>
-      {image && (
-        <>
-          <div className="modal-background">
-            <span className="close" onClick={() => setImage(null)}>
-              &times;
-            </span>
-            <img alt="" src={image} className="modal-content" />
-            {caption && (
-              <div className="caption">
-                <h2>{caption}</h2>
-              </div>
-            )}
-          </div>
-        </>
-      )}
-      <ImageGallery images={LegoImages.reverse()} openModal={openModal} />
+      <ImageGallery images={LegoImages.reverse()} />
     </div>
   );
 }
